Load initial form data on server
