+++ next/frontend/types/initialFormData.ts
@@ -1,0 +1,7 @@
+import { GetFileResponseDto } from '@clients/openapi-forms'
+
+export type InitialFormData = {
+  formDataJson: object
+  formId: string
+  files: GetFileResponseDto[]
+}
